Added support for review payment scaling.

+++ clients/cronos/applications/online_review/branches/reviewer_payment_scaling_20081015/src/java/main/com/cronos/onlinereview/phases/AutoPaymentUtil.java
@@ -125,8 +125,8 @@ public class AutoPaymentUtil {
         int passedCount = getScreenPassedCount(projectId, conn);
         float[] payments = getPayments(projectId, projectCategoryId, conn);
 
-        float prize = getPriceByProjectId(projectId, conn);
-        float drPoints = getDrPointsByProjectId(projectId, conn);
+        float prize = (float) getPriceByProjectId(projectId, conn);
+        float drPoints = (float) getDrPointsByProjectId(projectId, conn);
 
         FixedPriceComponent fpc = new FixedPriceComponent(levelId, count, passedCount,
                                                           (int) (projectCategoryId + 111),
